requirements.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
Bottleneck==1.3.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorlog==6.8.2
contourpy==1.2.0
cycler==0.11.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.25.0
fsspec==2023.10.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
keras==3.1.0
keras-tuner==1.4.7
kiwisolver==1.4.4
kt-legacy==1.0.5
libclang==18.1.1
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
munkres==1.1.4
namex==0.0.7
networkx==3.1
numexpr==2.8.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.10.0
optuna==3.6.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pip==23.3.1
protobuf==4.25.3
psutil==5.9.8
pycparser==2.21
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorchtools==0.0.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.12.2
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.28
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.3.0
tornado==6.3.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
